requirements.txt
fastapi
uvicorn
supabase
python-dotenv
gradio
groq
matplotlib

